backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15
django-cors-headers>=4.3
requests>=2.31
python-dotenv>=1.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
gunicorn>=22.0.0
